+++ app/uk/gov/hmrc/hecapplicantfrontend/models/RetrievedJourneyData.scala
@@ -17,6 +17,8 @@
 package uk.gov.hmrc.hecapplicantfrontend.models
 
 import play.api.libs.json._
+import uk.gov.hmrc.hecapplicantfrontend.models.hecTaxCheck.company.CTStatusResponse
+import uk.gov.hmrc.hecapplicantfrontend.models.hecTaxCheck.individual.SAStatusResponse
 import uk.gov.hmrc.hecapplicantfrontend.models.ids.CTUTR
 
 sealed trait RetrievedJourneyData extends Product with Serializable
